[WEAV-337] Locations 정보 Splash -> Main으로 호출 시점 변경
